+++ invitation/app/resources/views/invitations/receipt.tpl.php
